Allows for finer-grained control of secrets
TL;DR
-----

Updates `secrets` scripts so it's no longer all or nothing

Details
-------

After working with the `secrets` script for a bit it's become
apparent that the all-or-nothing approach was easy but not ideal.
The primary challenge is that encrypting a changex secret re-
encrypted all secrets which created a spurious change for all of
the other secrets.

This change changes the default mode to encrypting/decrypting a
specific secret file vs. all. This will avoid the spurious changes
and show a truer histroy of secret change.

It's still possible to use the previous behavior by specifying
the `--all` flag.